Separate Left panel and move QtWaveModel to single file
